+++ python/binding/bind_occupancy_octree_node.cpp
@@ -5,12 +5,5 @@
 void
 BindOccupancyOctreeNode(const py::module& m) {
     using namespace erl::geometry;
-    BindOccupancyOctreeNode<OccupancyOctreeNode, AbstractOctreeNode>(m, "OccupancyOctreeNode")
-        .def_property_readonly("occupancy", &OccupancyOctreeNode::GetOccupancy)
-        .def_property_readonly("log_odds", &OccupancyOctreeNode::GetLogOdds)
-        .def_property_readonly("mean_child_log_odds", &OccupancyOctreeNode::GetMeanChildLogOdds)
-        .def_property_readonly("max_child_log_odds", &OccupancyOctreeNode::GetMaxChildLogOdds)
-        .def("allow_update_log_odds", &OccupancyOctreeNode::AllowUpdateLogOdds, py::arg("delta"))
-        .def("add_log_odds", &OccupancyOctreeNode::AddLogOdds, py::arg("log_odds"))
-        .def("get_child", py::overload_cast<uint32_t>(&OccupancyOctreeNode::GetChild<OccupancyOctreeNode>, py::const_), py::arg("child_idx"));
+    BindOccupancyOctreeNode<OccupancyOctreeNode, AbstractOctreeNode>(m, "OccupancyOctreeNode");
 }