+++ python/binding/bind_pyobject_occupancy_octree.cpp
@@ -1,0 +1,10 @@
+#include "pyobject_occupancy_octree.hpp"
+
+#include "erl_common/pybind11.hpp"
+#include "erl_geometry/pybind11_occupancy_octree.hpp"
+
+void
+BindPyObjectOccupancyOctree(const py::module &m) {
+    using namespace erl::geometry;
+    BindOccupancyOctree<PyObjectOccupancyOctree, PyObjectOccupancyOctreeNode>(m, "PyObjectOccupancyOctree");
+}